+++ js/data.js
@@ -1,0 +1,47 @@
+const TITLES = [
+  'Хорошее жилье',
+  'Прекрасный вариант',
+  'Лучше не найдешь',
+  'Так себе жилье',
+  'За то не дорого',
+  'Милая маленькая квартирка',
+  'Шикарный вариант',
+  'Красиво жить не запретишь',
+  'Один раз живем',
+];
+const TYPE = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+  'hotel',
+];
+const TIMES = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+const FEATURES = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner',
+];
+const DESTCRIPTIONS = [
+  'Просторная светлая квартира, укоплектованна мебелью',
+  'Квартира расположенна на верхнем этаже, роскошный вид на город',
+  'Уютная квартира с необходимым комлектом мебели, с оборудованной кухней',
+  'Просторная квартира с панорамными окнами и чудесным видом из окна',
+  'Малогабаритная квартира для маленькой компании',
+  'Удобная квартира, укомлектованная всем необходимым для путешественников',
+];
+const PHOTOS = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
+];
+
+
+export {TITLES, TYPE, TIMES, FEATURES, DESTCRIPTIONS, PHOTOS};
